Extract `theme` into its own crate
